Extracted test cases for 'Vectors' strategy into separate modules
